backend/requirements.txt
# Django Core
Django>=5.0,<5.1
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-cors-headers>=4.3,<5.0
django-filter>=24.1,<25.0
drf-spectacular>=0.27,<1.0

# Environment & Config
django-environ>=0.11,<1.0

# Database
psycopg2-binary>=2.9,<3.0

# Geospatial (PostGIS)
# django.contrib.gis included in Django

# Cache & Message Broker
django-redis>=5.4,<6.0

# Async Tasks
celery>=5.3,<6.0

# WebSocket / Real-time
channels>=4.0,<5.0
channels-redis>=4.2,<5.0
daphne>=4.1,<5.0

# Media & Storage
Pillow>=10.2,<11.0
django-cloudinary-storage>=0.3,<1.0
cloudinary>=1.38,<2.0

# Payments
stripe>=8.0,<10.0

# Server
gunicorn>=21.2,<23.0
whitenoise>=6.6,<7.0

# Admin extras
django-import-export>=4.0,<5.0

# Security
argon2-cffi>=23.1,<24.0

# Testing
pytest>=8.0,<9.0
pytest-django>=4.8,<5.0
factory-boy>=3.3,<4.0
pytest-asyncio>=0.23,<1.0

# Linting
black>=24.0,<25.0
isort>=5.13,<6.0
